requirements.txt
aiohttp==3.6.2
appnope==0.1.0
async-timeout==3.0.1
attrs==19.3.0
backcall==0.2.0
backtrader==1.9.76.123
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
cycler==0.10.0
dateutils==0.6.8
decorator==4.4.2
defusedxml==0.6.0
dnspython==2.0.0
email-validator==1.1.1
entrypoints==0.3
ffn==0.3.4
filelock==3.0.12
future==0.18.2
google==3.0.0
grpcio==1.30.0
gym==0.17.2
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
llvmlite==0.33.0
Logbook==1.5.3
lxml==4.5.2
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
modin==0.7.4
more-itertools==8.4.0
msgpack==1.0.0
multidict==4.7.6
multitasking==0.0.9
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numba==0.50.1
numpy==1.19.1
packaging==20.4
pandas-datareader==0.9.0
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
py-spy==0.3.3
pyarrow==0.16.0
pyglet==1.5.0
Pygments==2.6.1
pyparsing==2.4.7
PyPrind==2.11.2
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
QtPy==1.9.0
Quandl==3.5.2
ray==0.8.6
redis==3.4.1
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.2
selenium==3.141.0
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
TA-Lib==0.4.18
tabulate==0.8.7
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
urllib3==1.25.10
webencodings==0.5.1
webull==0.1.0
widgetsnbextension==3.5.1
yarl==1.4.2
yfinance==0.1.54
zipp==3.1.0
